feat: Display debt notifications in header
Introduces a notification dropdown in the header to alert users about upcoming or overdue debts.

Fetches debt data and filters for debts that are overdue or due within the next 7 days. This data is then displayed using a new `BellIcon` and an associated dropdown menu in the application header.

Also includes minor updates to chart initialization and icon imports.